requirements.txt
PyQt==5.15.5
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0